+++ internal/app/storage.go
@@ -19,8 +19,6 @@ import (
 	"github.com/otterscale/otterscale/internal/core/storage/smb"
 )
 
-const DefaultSMBNamespace = "samba-operator-system"
-
 type StorageService struct {
 	pbconnect.UnimplementedStorageServiceHandler
 
@@ -452,7 +450,7 @@ func (s *StorageService) DeleteUserKey(ctx context.Context, req *pb.DeleteUserKe
 }
 
 func (s *StorageService) ListSMBShares(ctx context.Context, req *pb.ListSMBSharesRequest) (*pb.ListSMBSharesResponse, error) {
-	shares, hostname, err := s.smb.ListSMBShares(ctx, req.GetScope(), DefaultSMBNamespace)
+	shares, hostname, err := s.smb.ListSMBShares(ctx, req.GetScope())
 	if err != nil {
 		return nil, err
 	}
@@ -486,7 +484,6 @@ func (s *StorageService) CreateSMBShare(ctx context.Context, req *pb.CreateSMBSh
 
 	share, hostname, err := s.smb.CreateSMBShare(ctx,
 		req.GetScope(),
-		DefaultSMBNamespace,
 		req.GetName(),
 		req.GetSizeBytes(),
 		req.GetPort(),
@@ -531,7 +528,6 @@ func (s *StorageService) UpdateSMBShare(ctx context.Context, req *pb.UpdateSMBSh
 
 	share, hostname, err := s.smb.UpdateSMBShare(ctx,
 		req.GetScope(),
-		DefaultSMBNamespace,
 		req.GetName(),
 		req.GetSizeBytes(),
 		req.GetPort(),
@@ -553,7 +549,7 @@ func (s *StorageService) UpdateSMBShare(ctx context.Context, req *pb.UpdateSMBSh
 }
 
 func (s *StorageService) ValidateSMBUser(ctx context.Context, req *pb.ValidateSMBUserRequest) (*pb.ValidateSMBUserResponse, error) {
-	result, err := s.smb.ADValidate(ctx,
+	result, err := s.smb.ValidateSMBUser(ctx,
 		req.GetRealm(),
 		req.GetUsername(),
 		req.GetPassword(),
@@ -567,7 +563,6 @@ func (s *StorageService) ValidateSMBUser(ctx context.Context, req *pb.ValidateSM
 	resp.SetValid(result.Valid)
 	resp.SetEntityType(toProtoEntityType(result.EntityType))
 	resp.SetMessage(result.Message)
-
 	return resp, nil
 }
 
@@ -1064,7 +1059,6 @@ func toProtoSMBShare(sd *smb.ShareData, hostname string) *pb.SMBShare {
 	share := sd.Share
 	if share != nil {
 		ret.SetName(share.Name)
-		ret.SetNamespace(share.Namespace)
 
 		service := sd.Service
 		if service != nil {
